Add filetype handling for jsonschema and update plugins for nvim-cmp
This commit introduces automatic filetype recognition for `*.jsonschema` files as JSON and overhauls the Neovim configuration to integrate the nvim-cmp auto-completion plugin, replacing the previous NCM2 setup. Additionally, it adds the configuration for lspconfig, which sets up language servers for YAML, JSON, and Terraform with the new completion capabilities. Finally, updates related to Vim-Plug handling of Lua files for Neovim are included.